move plugin loader to extension directory
