+++ Composer/packages/extensions/plugin-loader/.eslintrc.js
@@ -1,5 +1,5 @@
 module.exports = {
-  extends: ['../../.eslintrc.js'],
+  extends: ['../../../.eslintrc.js'],
   parserOptions: {
     project: './tsconfig.json',
     tsconfigRootDir: __dirname,
