+++ Composer/packages/extensions/plugin-loader/.gitignore
