+++ Composer/packages/extensions/plugin-loader/package.json
