+++ Composer/packages/extensions/plugin-loader/src/index.ts
