+++ Composer/packages/extensions/plugin-loader/src/logger.ts
