+++ Composer/packages/extensions/plugin-loader/src/types.ts
